fix: replace context with local state
